+++ boxsdk/__init__.py
@@ -2,6 +2,10 @@
 
 from __future__ import unicode_literals
 
+try:
+    from .auth.jwt_auth import JWTAuth
+except ImportError:
+    JWTAuth = None  # If extras are not installed, JWTAuth won't be available.
 from .auth.oauth2 import OAuth2
 from .client import Client
 from .object import *  # pylint:disable=wildcard-import,redefined-builtin