+++ HISTORY.rst
@@ -3,6 +3,37 @@
 Release History
 ---------------
 
+Upcoming
+++++++++
+
+- CPython 3.5 support.
+- Support for cryptography>=1.0 on PyPy 2.6.
+- Travis CI testing for CPython 3.5 and PyPy 2.6.0.
+
+1.2.1 (2015-07-22)
+++++++++++++++++++
+
+- The SDK now supports setting a password when creating a shared link.
+
+1.2.1 (2015-07-22)
+++++++++++++++++++
+
+**Bugfixes**
+
+- Fixed an ImportError for installs that didn't install the [jwt] extras.
+
+1.2.0 (2015-07-13)
+++++++++++++++++++
+
+- Added support for Box Developer Edition. This includes JWT auth (auth as enterprise or as app user),
+  and `create_user` functionality.
+- Added support for setting shared link expiration dates.
+- Added support for setting shared link permissions.
+- Added support for 'As-User' requests. See https://box-content.readme.io/#as-user-1
+- Improved support for accessing shared items. Items returned from the `client.get_shared_item` method will
+  remember the shared link (and the optionally provided shared link password) so methods called on the returned
+  items will be properly authorized.
+
 1.1.7 (2015-05-28)
 ++++++++++++++++++
 
@@ -12,7 +43,7 @@ Release History
 
 - ``Item.remove_shared_link()`` was trying to return an incorrect (according to its own documentation) value, and was
   also attempting to calculate that value in a way that made an incorrect assumption about the API response. The latter
-  problem caused invokations of the method to raise TypeError. The method now handles the response correctly, and
+  problem caused invocations of the method to raise TypeError. The method now handles the response correctly, and
   correctly returns type ``bool``.
 
 1.1.6 (2015-04-17)